Refactored Ramping VUs executor—_a little bit_. This is the first step as I'm learning about the system. I hope this refactoring makes sense.