@na-- 2021-08-30
Thanks, I just skimmed this and it seems like a solid improvement! That said, all of the new methods end up having a ton of parameters and results, which is not ideal for comprehension. What we ended up doing in another executor (the `externally-controlled`) one was to split up the `Run()` state into a separate `struct` and have all of these things as methods to that `struct`. Take a look at https://github.com/grafana/k6/blob/3994f1a5845ebf5d8a22d2fd11f1ae5c25761ae3/lib/executor/externally_controlled.go#L373-L387