@inancgumus 2021-08-30
Thanks! Also, thanks for pointing out the `externallyControlledRunState` implementation :) It helped.

I have added a struct called `rampingVUsRunState` to share data between the methods. I still don't think it's an ideal implementation, but I hope it paves the way for the next possible refactoring steps.