@MStoykov 2021-09-28
@inancgumus can you also provide some benchmark results from before and after the refactor