@inancgumus 2021-09-28
@MStoykov 

```
name                                                                                                      old time/op    new time/op    delta
RampingVUsGetRawExecutionSteps/seq:;segment:/normal-8                                                        221µs ± 0%     221µs ± 0%   ~     (p=1.000 n=1+1)
RampingVUsGetRawExecutionSteps/seq:;segment:/rollercoaster-8                                                2.54ms ± 0%    2.40ms ± 0%   ~     (p=1.000 n=1+1)
RampingVUsGetRawExecutionSteps/seq:;segment:0:1/normal-8                                                     221µs ± 0%     219µs ± 0%   ~     (p=1.000 n=1+1)
RampingVUsGetRawExecutionSteps/seq:;segment:0:1/rollercoaster-8                                             2.44ms ± 0%    2.45ms ± 0%   ~     (p=1.000 n=1+1)
RampingVUsGetRawExecutionSteps/seq:0,0.3,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.3/normal-8                       66.3µs ± 0%    66.8µs ± 0%   ~     (p=1.000 n=1+1)
RampingVUsGetRawExecutionSteps/seq:0,0.3,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.3/rollercoaster-8                 739µs ± 0%     746µs ± 0%   ~     (p=1.000 n=1+1)
RampingVUsGetRawExecutionSteps/seq:0,0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.1/normal-8           22.9µs ± 0%    23.0µs ± 0%   ~     (p=1.000 n=1+1)
RampingVUsGetRawExecutionSteps/seq:0,0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.1/rollercoaster-8     241µs ± 0%     247µs ± 0%   ~     (p=1.000 n=1+1)
RampingVUsGetRawExecutionSteps/seq:;segment:2/5:4/5/normal-8                                                89.0µs ± 0%    90.9µs ± 0%   ~     (p=1.000 n=1+1)
RampingVUsGetRawExecutionSteps/seq:;segment:2/5:4/5/rollercoaster-8                                         1.02ms ± 0%    1.07ms ± 0%   ~     (p=1.000 n=1+1)
RampingVUsGetRawExecutionSteps/seq:;segment:2235/5213:4/5/normal-8                                          87.9µs ± 0%    89.6µs ± 0%   ~     (p=1.000 n=1+1)
RampingVUsGetRawExecutionSteps/seq:;segment:2235/5213:4/5/rollercoaster-8                                    911µs ± 0%     939µs ± 0%   ~     (p=1.000 n=1+1)

name                                                                                                      old alloc/op   new alloc/op   delta
RampingVUsGetRawExecutionSteps/seq:;segment:/normal-8                                                        246kB ± 0%     246kB ± 0%   ~     (p=1.000 n=1+1)
RampingVUsGetRawExecutionSteps/seq:;segment:/rollercoaster-8                                                2.65MB ± 0%    2.65MB ± 0%   ~     (p=1.000 n=1+1)
RampingVUsGetRawExecutionSteps/seq:;segment:0:1/normal-8                                                     246kB ± 0%     246kB ± 0%   ~     (all equal)
RampingVUsGetRawExecutionSteps/seq:;segment:0:1/rollercoaster-8                                             2.65MB ± 0%    2.65MB ± 0%   ~     (p=1.000 n=1+1)
RampingVUsGetRawExecutionSteps/seq:0,0.3,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.3/normal-8                       73.7kB ± 0%    73.7kB ± 0%   ~     (all equal)
RampingVUsGetRawExecutionSteps/seq:0,0.3,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.3/rollercoaster-8                 795kB ± 0%     795kB ± 0%   ~     (p=1.000 n=1+1)
RampingVUsGetRawExecutionSteps/seq:0,0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.1/normal-8           24.6kB ± 0%    24.6kB ± 0%   ~     (all equal)
RampingVUsGetRawExecutionSteps/seq:0,0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.1/rollercoaster-8     270kB ± 0%     270kB ± 0%   ~     (all equal)
RampingVUsGetRawExecutionSteps/seq:;segment:2/5:4/5/normal-8                                                98.3kB ± 0%    98.3kB ± 0%   ~     (all equal)
RampingVUsGetRawExecutionSteps/seq:;segment:2/5:4/5/rollercoaster-8                                         1.06MB ± 0%    1.06MB ± 0%   ~     (p=1.000 n=1+1)
RampingVUsGetRawExecutionSteps/seq:;segment:2235/5213:4/5/normal-8                                          90.1kB ± 0%    90.1kB ± 0%   ~     (all equal)
RampingVUsGetRawExecutionSteps/seq:;segment:2235/5213:4/5/rollercoaster-8                                    983kB ± 0%     983kB ± 0%   ~     (p=1.000 n=1+1)

name                                                                                                      old allocs/op  new allocs/op  delta
RampingVUsGetRawExecutionSteps/seq:;segment:/normal-8                                                         1.00 ± 0%      1.00 ± 0%   ~     (all equal)
RampingVUsGetRawExecutionSteps/seq:;segment:/rollercoaster-8                                                  1.00 ± 0%      1.00 ± 0%   ~     (all equal)
RampingVUsGetRawExecutionSteps/seq:;segment:0:1/normal-8                                                      1.00 ± 0%      1.00 ± 0%   ~     (all equal)
RampingVUsGetRawExecutionSteps/seq:;segment:0:1/rollercoaster-8                                               1.00 ± 0%      1.00 ± 0%   ~     (all equal)
RampingVUsGetRawExecutionSteps/seq:0,0.3,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.3/normal-8                         1.00 ± 0%      1.00 ± 0%   ~     (all equal)
RampingVUsGetRawExecutionSteps/seq:0,0.3,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.3/rollercoaster-8                  1.00 ± 0%      1.00 ± 0%   ~     (all equal)
RampingVUsGetRawExecutionSteps/seq:0,0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.1/normal-8             1.00 ± 0%      1.00 ± 0%   ~     (all equal)
RampingVUsGetRawExecutionSteps/seq:0,0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.1/rollercoaster-8      1.00 ± 0%      1.00 ± 0%   ~     (all equal)
RampingVUsGetRawExecutionSteps/seq:;segment:2/5:4/5/normal-8                                                  1.00 ± 0%      1.00 ± 0%   ~     (all equal)
RampingVUsGetRawExecutionSteps/seq:;segment:2/5:4/5/rollercoaster-8                                           1.00 ± 0%      1.00 ± 0%   ~     (all equal)
RampingVUsGetRawExecutionSteps/seq:;segment:2235/5213:4/5/normal-8                                            1.00 ± 0%      1.00 ± 0%   ~     (all equal)
RampingVUsGetRawExecutionSteps/seq:;segment:2235/5213:4/5/rollercoaster-8                                     1.00 ± 0%      1.00 ± 0%   ~     (all equal)
```