@inancgumus 2021-09-28
Yes, you're right, I forgot to do that. I was benchmarking just before our buddy session with @imiric :) And, I'll rebase tomorrow; it's alright.

Here are the proper benchmarks:

```
name                                                                                                      old time/op    new time/op    delta
RampingVUsGetRawExecutionSteps/seq:;segment:/normal-8                                                        223µs ± 5%     236µs ±16%    ~     (p=0.156 n=9+10)
RampingVUsGetRawExecutionSteps/seq:;segment:/rollercoaster-8                                                2.59ms ± 5%    2.58ms ± 2%    ~     (p=0.356 n=10+9)
RampingVUsGetRawExecutionSteps/seq:;segment:0:1/normal-8                                                     233µs ± 6%     226µs ±11%    ~     (p=0.165 n=10+10)
RampingVUsGetRawExecutionSteps/seq:;segment:0:1/rollercoaster-8                                             2.60ms ± 5%    2.57ms ± 8%    ~     (p=0.481 n=10+10)
RampingVUsGetRawExecutionSteps/seq:0,0.3,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.3/normal-8                       67.2µs ± 6%    65.0µs ± 1%  -3.20%  (p=0.015 n=10+10)
RampingVUsGetRawExecutionSteps/seq:0,0.3,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.3/rollercoaster-8                 738µs ± 5%     729µs ± 3%    ~     (p=0.280 n=10+10)
RampingVUsGetRawExecutionSteps/seq:0,0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.1/normal-8           24.1µs ±10%    21.9µs ± 1%  -9.06%  (p=0.000 n=10+8)
RampingVUsGetRawExecutionSteps/seq:0,0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.1/rollercoaster-8     241µs ± 2%     237µs ± 2%  -1.67%  (p=0.023 n=10+10)
RampingVUsGetRawExecutionSteps/seq:;segment:2/5:4/5/normal-8                                                87.2µs ± 3%    87.1µs ± 3%    ~     (p=0.393 n=10+10)
RampingVUsGetRawExecutionSteps/seq:;segment:2/5:4/5/rollercoaster-8                                          964µs ± 3%     963µs ± 2%    ~     (p=0.905 n=10+9)
RampingVUsGetRawExecutionSteps/seq:;segment:2235/5213:4/5/normal-8                                          85.2µs ± 1%    85.8µs ± 2%    ~     (p=0.278 n=9+10)
RampingVUsGetRawExecutionSteps/seq:;segment:2235/5213:4/5/rollercoaster-8                                    905µs ± 3%     896µs ± 0%    ~     (p=0.243 n=10+9)

name                                                                                                      old alloc/op   new alloc/op   delta
RampingVUsGetRawExecutionSteps/seq:;segment:/normal-8                                                        246kB ± 0%     246kB ± 0%    ~     (p=0.591 n=9+10)
RampingVUsGetRawExecutionSteps/seq:;segment:/rollercoaster-8                                                2.65MB ± 0%    2.65MB ± 0%    ~     (p=0.710 n=10+9)
RampingVUsGetRawExecutionSteps/seq:;segment:0:1/normal-8                                                     246kB ± 0%     246kB ± 0%    ~     (all equal)
RampingVUsGetRawExecutionSteps/seq:;segment:0:1/rollercoaster-8                                             2.65MB ± 0%    2.65MB ± 0%    ~     (p=0.130 n=8+10)
RampingVUsGetRawExecutionSteps/seq:0,0.3,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.3/normal-8                       73.7kB ± 0%    73.7kB ± 0%    ~     (all equal)
RampingVUsGetRawExecutionSteps/seq:0,0.3,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.3/rollercoaster-8                 795kB ± 0%     795kB ± 0%    ~     (p=0.058 n=7+9)
RampingVUsGetRawExecutionSteps/seq:0,0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.1/normal-8           24.6kB ± 0%    24.6kB ± 0%    ~     (all equal)
RampingVUsGetRawExecutionSteps/seq:0,0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.1/rollercoaster-8     270kB ± 0%     270kB ± 0%    ~     (all equal)
RampingVUsGetRawExecutionSteps/seq:;segment:2/5:4/5/normal-8                                                98.3kB ± 0%    98.3kB ± 0%    ~     (all equal)
RampingVUsGetRawExecutionSteps/seq:;segment:2/5:4/5/rollercoaster-8                                         1.06MB ± 0%    1.06MB ± 0%    ~     (p=1.000 n=10+10)
RampingVUsGetRawExecutionSteps/seq:;segment:2235/5213:4/5/normal-8                                          90.1kB ± 0%    90.1kB ± 0%    ~     (all equal)
RampingVUsGetRawExecutionSteps/seq:;segment:2235/5213:4/5/rollercoaster-8                                    983kB ± 0%     983kB ± 0%    ~     (p=0.124 n=10+10)

name                                                                                                      old allocs/op  new allocs/op  delta
RampingVUsGetRawExecutionSteps/seq:;segment:/normal-8                                                         1.00 ± 0%      1.00 ± 0%    ~     (all equal)
RampingVUsGetRawExecutionSteps/seq:;segment:/rollercoaster-8                                                  1.00 ± 0%      1.00 ± 0%    ~     (all equal)
RampingVUsGetRawExecutionSteps/seq:;segment:0:1/normal-8                                                      1.00 ± 0%      1.00 ± 0%    ~     (all equal)
RampingVUsGetRawExecutionSteps/seq:;segment:0:1/rollercoaster-8                                               1.00 ± 0%      1.00 ± 0%    ~     (all equal)
RampingVUsGetRawExecutionSteps/seq:0,0.3,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.3/normal-8                         1.00 ± 0%      1.00 ± 0%    ~     (all equal)
RampingVUsGetRawExecutionSteps/seq:0,0.3,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.3/rollercoaster-8                  1.00 ± 0%      1.00 ± 0%    ~     (all equal)
RampingVUsGetRawExecutionSteps/seq:0,0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.1/normal-8             1.00 ± 0%      1.00 ± 0%    ~     (all equal)
RampingVUsGetRawExecutionSteps/seq:0,0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.1/rollercoaster-8      1.00 ± 0%      1.00 ± 0%    ~     (all equal)
RampingVUsGetRawExecutionSteps/seq:;segment:2/5:4/5/normal-8                                                  1.00 ± 0%      1.00 ± 0%    ~     (all equal)
RampingVUsGetRawExecutionSteps/seq:;segment:2/5:4/5/rollercoaster-8                                           1.00 ± 0%      1.00 ± 0%    ~     (all equal)
RampingVUsGetRawExecutionSteps/seq:;segment:2235/5213:4/5/normal-8                                            1.00 ± 0%      1.00 ± 0%    ~     (all equal)
RampingVUsGetRawExecutionSteps/seq:;segment:2235/5213:4/5/rollercoaster-8                                     1.00 ± 0%      1.00 ± 0%    ~     (all equal)

```